# Improve TimeSpan.ToString/TryFormat throughput for default format